Move nbhide extension to subdirectory, make setup file
