+++ nbhide/ext/MANIFEST.in
