+++ nbhide/ext/jupyter-config/jupyter_notebook_config.d/nbhide.json
