+++ nbhide/ext/jupyter-config/nbconfig/notebook.d/nbhide.json
