+++ nbhide/ext/nbhide/static/appendButtons.js
