+++ nbhide/ext/nbhide/static/appendToggles.js
