+++ nbhide/setup.py
@@ -1,7 +1,12 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-"""
-Adds toolbar controls for (un)hiding of code and Markdown cells
+
+"""Minimal code cell hiding using HTML styles and D3.js
+
+   Eric Easthope
+   
+   MIT License
+   Assembled for Callysto
 """
 
 from setuptools import setup, find_packages
@@ -9,32 +14,11 @@ from setuptools import setup, find_packages
 def main():
     setup(
         name='nbhide',
-        version='0.0.dev0',
-        author='Vincent Cote & Eric Easthope',
-        packages=find_packages(),
+        version='0.0.0',
+        author='Eric Easthope',
+        packages=find_packages(exclude=['ext']),
         include_package_data=True,
-        
-        data_files=[
-            # like `jupyter nbextension install --sys-prefix`
-            ("share/jupyter/nbextensions/nbhide", [
-                "nbhide/static/appendButtons.js",
-            ]),
-            
-            # like `jupyter nbextension install --sys-prefix`
-            ("share/jupyter/nbextensions/nbhide", [
-                "nbhide/static/appendToggles.js",
-            ]),
-            
-            # like `jupyter nbextension enable --sys-prefix`
-            ("etc/jupyter/nbconfig/notebook.d", [
-                "jupyter-config/nbconfig/notebook.d/nbhide.json"
-            ]),
-            
-            # like `jupyter serverextension enable --sys-prefix`
-            ("etc/jupyter/jupyter_notebook_config.d", [
-                "jupyter-config/jupyter_notebook_config.d/nbhide.json"
-            ])
-        ]
+        install_requires=[]
     )
 
 if __name__ == '__main__':
